requirements.txt
torch==1.0.0
numpy==1.16.2
torchvision==0.2.1
Pillow==6.0.0
